+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Brackets Todo Changelog
 
-## 0.2
+## 0.3.0
+* Scope of search can be set to currently open document or entire project.
+* Option for case sensitive searches.
+
+## 0.2.0
 * Default settings, including regular expression used and tags, can be customized by adding a .todo file in your project root.
 * Colon after tags are now optional to allow easier customization of comment syntax.
 * TODO, FIXME, NOTE and CHANGES tags have colors for an improved experience and easier navigation. Thanks to António Pinto (apintocr).
 
-## 0.1
+## 0.1.2
 * Initial release.